+++ lilith/src/main/java/de/huxhorn/lilith/swing/actions/FocusHttpRequestHeaderAction.java
@@ -46,6 +46,10 @@ public class FocusHttpRequestHeaderAction
 		{
 			return null;
 		}
+		if(isAlternativeBehaviorRequested(e))
+		{
+			return new HttpRequestHeadersContainsCondition(key, null);
+		}
 		return new HttpRequestHeadersContainsCondition(key, value);
 	}
 }