+++ lilith/src/main/java/de/huxhorn/lilith/swing/actions/FocusHttpResponseHeaderAction.java
@@ -46,6 +46,10 @@ public class FocusHttpResponseHeaderAction
 		{
 			return null;
 		}
+		if(isAlternativeBehaviorRequested(e))
+		{
+			return new HttpResponseHeadersContainsCondition(key, null);
+		}
 		return new HttpResponseHeadersContainsCondition(key, value);
 	}
 }