+++ lilith/src/main/java/de/huxhorn/lilith/swing/menu/ExcludeHttpStatusTypeMenu.java
@@ -19,7 +19,6 @@ package de.huxhorn.lilith.swing.menu;
 
 import de.huxhorn.lilith.data.access.HttpStatus;
 import de.huxhorn.lilith.swing.actions.BasicFilterAction;
-import de.huxhorn.lilith.swing.actions.FocusHttpStatusTypeAction;
 import de.huxhorn.lilith.swing.actions.NegateFilterAction;
 
 class ExcludeHttpStatusTypeMenu
@@ -27,8 +26,9 @@ class ExcludeHttpStatusTypeMenu
 {
 	private static final long serialVersionUID = 4925999851649779333L;
 
+	@Override
 	protected BasicFilterAction createAction(HttpStatus.Type type)
 	{
-		return new NegateFilterAction(new FocusHttpStatusTypeAction(type));
+		return new NegateFilterAction(super.createAction(type));
 	}
 }