+++ lilith/src/main/java/de/huxhorn/lilith/swing/menu/ExcludeNDCMenu.java
@@ -18,8 +18,6 @@
 package de.huxhorn.lilith.swing.menu;
 
 import de.huxhorn.lilith.swing.actions.BasicFilterAction;
-import de.huxhorn.lilith.swing.actions.FocusNDCAction;
-import de.huxhorn.lilith.swing.actions.FocusNDCPatternAction;
 import de.huxhorn.lilith.swing.actions.NegateFilterAction;
 
 class ExcludeNDCMenu
@@ -35,12 +33,12 @@ class ExcludeNDCMenu
 	@Override
 	protected BasicFilterAction createMessageAction(String message)
 	{
-		return new NegateFilterAction(new FocusNDCAction(message, htmlTooltip));
+		return new NegateFilterAction(super.createMessageAction(message));
 	}
 
 	@Override
 	protected BasicFilterAction createMessagePatternAction(String pattern)
 	{
-		return new NegateFilterAction(new FocusNDCPatternAction(pattern, htmlTooltip));
+		return new NegateFilterAction(super.createMessagePatternAction(pattern));
 	}
 }